requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
fake-useragent==0.1.11
httmock==1.4.0
http-request-randomizer==1.3.2
idna==2.10
numpy==1.20.1
pandas==1.2.2
psutil==5.8.0
pycparser==2.20
pyOpenSSL==20.0.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
selenium==3.141.0
six==1.15.0
soupsieve==2.2
urllib3==1.26.3
